+++ ICD.Connect.Conferencing.Zoom/EventArguments/MeetingNeedsPasswordEventArgs.cs
@@ -5,10 +5,13 @@ namespace ICD.Connect.Conferencing.Zoom.EventArguments
 	public sealed class MeetingNeedsPasswordEventArgs : EventArgs
 	{
 		private readonly string m_MeetingNumber;
+		private readonly string m_MeetingNumberFormatted;
 		private readonly bool m_NeedsPassword;
 		private readonly bool m_WrongAndRetry;
 
 		public string MeetingNumber { get { return m_MeetingNumber; } }
+
+		public string MeetingNumberFormatted { get { return m_MeetingNumberFormatted; } }
 
 		public bool NeedsPassword { get { return m_NeedsPassword; } }
 
@@ -18,11 +21,14 @@ namespace ICD.Connect.Conferencing.Zoom.EventArguments
 		/// Constructor.
 		/// </summary>
 		/// <param name="meetingNumber"></param>
+		/// <param name="meetingNumberFormatted"></param>
 		/// <param name="needsPassword"></param>
 		/// <param name="wrongAndRetry"></param>
-		public MeetingNeedsPasswordEventArgs(string meetingNumber, bool needsPassword, bool wrongAndRetry)
+		public MeetingNeedsPasswordEventArgs(string meetingNumber, string meetingNumberFormatted,
+		                                     bool needsPassword, bool wrongAndRetry)
 		{
 			m_MeetingNumber = meetingNumber;
+			m_MeetingNumberFormatted = meetingNumberFormatted;
 			m_NeedsPassword = needsPassword;
 			m_WrongAndRetry = wrongAndRetry;
 		}